Fix `ChangeWalletSheet` circular dependencies
