+++ src/screens/change-wallet/layout.ts
@@ -1,0 +1,6 @@
+import { DEVICE_HEIGHT } from '@/utils/deviceUtils';
+
+export const PANEL_INSET_HORIZONTAL = 20;
+export const MAX_PANEL_HEIGHT = Math.min(690, DEVICE_HEIGHT - 100);
+export const PANEL_HEADER_HEIGHT = 58;
+export const FOOTER_HEIGHT = 91;
